cut details

+++ src/components/cuts/ShowCutsTable.tsx
@@ -4,6 +4,9 @@ import { Button, Preset } from "../button/button";
 import useConfigStore from "@/stores/configStore";
 import { numberToMoney } from "@/lib/utils";
 import { Cut } from "@/interfaces/Cuts";
+import { DeleteCutButton } from "./DeleteCutsButton";
+import { CutDetailsModal } from "./CutDetailsModal";
+import useModalStore from "@/stores/modalStorage";
 
 
 export interface ShowCutsTableProps {
@@ -13,22 +16,25 @@ export interface ShowCutsTableProps {
 export function ShowCutsTable(props: ShowCutsTableProps) {
   const { records } = props;
   const { system } = useConfigStore();
+  const { modals, openModal, closeModal } = useModalStore();
 
-const firstRecord = records && records[0];
-
+  const firstRecord = records && records[0];
+  if (!records || records.length == 0) {
+    return null
+  }
 
   const listItems = records &&  records.map((record: any, key: any) => (
     <tr key={key} className={`border-2 ${record.status == 0 && "bg-red-100"}`} >
       <td className="py-2 px-6 truncate clickeable" onClick={()=>{}}>
         { record?.close && formatDateAsDMY(record.close) }  { record?.close ? formatTime(record.close) : "Sin corte"}
       </td>
-      <td className="py-2 px-6 clickeable" onClick={()=>{}}>
+      <td className="py-2 px-6 clickeable" onClick={()=>openModal('cutDetails')}>
         { record?.employee?.name }
       </td>
       <td className={`py-2 px-6 font-bold clickeable ${record?.cash_diference > 0 ? 'text-blue-600' : record?.cash_diference < 0 ? 'text-red-600' : 'text-black'}`} onClick={()=>{}}>{ numberToMoney(record?.cash_diference ?? 0, system) }
       </td>
       <td className="py-2 px-6">
-        <Button preset={firstRecord.id == record?.id ? Preset.smallClose : Preset.smallCloseDisable} onClick={ ()=>{} } noText />
+        <DeleteCutButton cut={record} isInitial={firstRecord.id == record?.id && record.status == 2} />
       </td>
     </tr>
   ));
@@ -47,5 +53,6 @@ const firstRecord = records && records[0];
                 <tbody>{listItems}</tbody>
             </table>
         </div>
+        <CutDetailsModal isShow={modals['cutDetails']} onClose={() => closeModal('cutDetails')} record={records[0]} />
    </div>);
 }
